Remove Unlocked vs Locked Wallet distinction
