+++ docs/wallet/wallet.md
@@ -62,7 +62,7 @@ import org.bitcoins.keymanager.bip39._
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
 import org.bitcoins.rpc.config.BitcoindInstance
 import org.bitcoins.wallet.config.WalletAppConfig
-import org.bitcoins.wallet.api.LockedWalletApi
+import org.bitcoins.wallet.api.WalletApi
 import org.bitcoins.wallet.Wallet
 
 import com.typesafe.config.ConfigFactory
@@ -151,7 +151,7 @@ val wallet = Wallet(keyManager, new NodeApi {
     override def getHeightByBlockStamp(blockStamp: BlockStamp): Future[Int] = Future.successful(0)
     override def getFiltersBetweenHeights(startHeight: Int, endHeight: Int): Future[Vector[FilterResponse]] = Future.successful(Vector.empty)
   })
-val walletF: Future[LockedWalletApi] = configF.flatMap { _ =>
+val walletF: Future[WalletApi] = configF.flatMap { _ =>
   Wallet.initialize(wallet,bip39PasswordOpt)
 }
