+++ wallet-test/src/test/scala/org/bitcoins/wallet/SegwitWalletTest.scala
@@ -3,18 +3,18 @@ package org.bitcoins.wallet
 import org.bitcoins.core.hd.{AddressType, HDPurposes}
 import org.bitcoins.core.protocol.{Bech32Address, P2PKHAddress, P2SHAddress}
 import org.bitcoins.testkit.wallet.BitcoinSWalletTest
-import org.bitcoins.wallet.api.UnlockedWalletApi
+import org.bitcoins.wallet.api.WalletApi
 import org.scalatest.FutureOutcome
 
 class SegwitWalletTest extends BitcoinSWalletTest {
 
-  override type FixtureParam = UnlockedWalletApi
+  override type FixtureParam = WalletApi
 
   override def withFixture(test: OneArgAsyncTest): FutureOutcome = {
     withSegwitWallet(test)
   }
 
-  it should "generate segwit addresses" in { wallet: UnlockedWalletApi =>
+  it should "generate segwit addresses" in { wallet: WalletApi =>
     for {
       addr <- wallet.getNewAddress()
       account <- wallet.getDefaultAccount()
