+++ wallet/src/main/scala/org/bitcoins/wallet/Wallet.scala
@@ -1,33 +1,205 @@
 package org.bitcoins.wallet
 
 import org.bitcoins.core.api.{ChainQueryApi, NodeApi}
+import org.bitcoins.core.bloom.{BloomFilter, BloomUpdateAll}
 import org.bitcoins.core.crypto._
 import org.bitcoins.core.currency._
-import org.bitcoins.core.hd._
+import org.bitcoins.core.hd.{HDAccount, HDCoin, HDPurposes}
 import org.bitcoins.core.protocol.BitcoinAddress
-import org.bitcoins.core.protocol.transaction._
+import org.bitcoins.core.protocol.blockchain.BlockHeader
+import org.bitcoins.core.protocol.transaction.{
+  Transaction,
+  TransactionOutPoint,
+  TransactionOutput
+}
 import org.bitcoins.core.wallet.fee.FeeUnit
+import org.bitcoins.core.wallet.utxo.TxoState
+import org.bitcoins.core.wallet.utxo.TxoState.{
+  ConfirmedReceived,
+  PendingConfirmationsReceived
+}
 import org.bitcoins.keymanager.KeyManagerParams
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
 import org.bitcoins.keymanager.util.HDUtil
 import org.bitcoins.wallet.api._
 import org.bitcoins.wallet.config.WalletAppConfig
-import org.bitcoins.wallet.models._
+import org.bitcoins.wallet.internal._
+import org.bitcoins.wallet.models.{SpendingInfoDb, _}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
 
-sealed abstract class Wallet extends LockedWallet with UnlockedWalletApi {
+abstract class Wallet
+    extends WalletApi
+    with UtxoHandling
+    with AddressHandling
+    with AccountHandling
+    with FundTransactionHandling
+    with TransactionProcessing
+    with RescanHandling {
 
-  val nodeApi: NodeApi
-  val chainQueryApi: ChainQueryApi
+  private[wallet] val addressDAO: AddressDAO = AddressDAO()
+  private[wallet] val accountDAO: AccountDAO = AccountDAO()
+  private[wallet] val spendingInfoDAO: SpendingInfoDAO = SpendingInfoDAO()
+  private[wallet] val transactionDAO: TransactionDAO = TransactionDAO()
+  private[wallet] val incomingTxDAO: IncomingTransactionDAO =
+    IncomingTransactionDAO()
+  private[wallet] val outgoingTxDAO: OutgoingTransactionDAO =
+    OutgoingTransactionDAO()
 
-  /**
-    * @inheritdoc
-    */
-  override def lock(): LockedWalletApi = {
-    logger.debug(s"Locking wallet")
-    LockedWallet(nodeApi, chainQueryApi)
+  override def isEmpty(): Future[Boolean] =
+    for {
+      addressCount <- addressDAO.count()
+      spendingInfoCount <- spendingInfoDAO.count()
+    } yield addressCount == 0 && spendingInfoCount == 0
+
+  override def clearUtxosAndAddresses(account: HDAccount): Future[Wallet] = {
+    for {
+      accountUtxos <- spendingInfoDAO.findAllForAccount(account)
+      deleteUtxoFs = accountUtxos.map(spendingInfoDAO.delete)
+      _ <- Future.sequence(deleteUtxoFs)
+      accountAddresses <- addressDAO.findAllForAccount(account)
+      deleteAddrFs = accountAddresses.map(addressDAO.delete)
+      _ <- Future.sequence(deleteAddrFs)
+    } yield this
+  }
+
+  override def clearAllUtxosAndAddresses(): Future[Wallet] = {
+    for {
+      _ <- spendingInfoDAO.deleteAll()
+      _ <- addressDAO.deleteAll()
+    } yield this
+  }
+
+  /** Sums up the value of all unspent
+    * TXOs in the wallet, filtered by the given predicate */
+  private def filterThenSum(
+      predicate: SpendingInfoDb => Boolean): Future[CurrencyUnit] = {
+    for (utxos <- spendingInfoDAO.findAllUnspentForAccount(
+           walletConfig.defaultAccount))
+      yield {
+        val filtered = utxos
+          .filter(predicate)
+          .map {
+            case txo: SpendingInfoDb =>
+              txo.state match {
+                case TxoState.PendingConfirmationsReceived |
+                    TxoState.ConfirmedReceived =>
+                  txo.output.value
+                case TxoState.Reserved | TxoState.PendingConfirmationsSpent |
+                    TxoState.ConfirmedSpent | TxoState.DoesNotExist =>
+                  CurrencyUnits.zero
+              }
+          }
+
+        filtered.fold(0.sats)(_ + _)
+      }
+  }
+
+  override def getConfirmedBalance(): Future[CurrencyUnit] = {
+    val confirmed = filterThenSum(_.state == ConfirmedReceived)
+    confirmed.foreach(balance =>
+      logger.trace(s"Confirmed balance=${balance.satoshis}"))
+    confirmed
+  }
+
+  override def getConfirmedBalance(account: HDAccount): Future[CurrencyUnit] = {
+    val allUnspentF = spendingInfoDAO.findAllUnspent()
+    val unspentInAccountF = for {
+      allUnspent <- allUnspentF
+    } yield {
+      allUnspent.filter { utxo =>
+        HDAccount.isSameAccount(utxo.privKeyPath.path, account) &&
+        utxo.blockHash.isDefined
+      }
+    }
+
+    unspentInAccountF.map(_.foldLeft(CurrencyUnits.zero)(_ + _.output.value))
+  }
+
+  override def getUnconfirmedBalance(): Future[CurrencyUnit] = {
+    val unconfirmed = filterThenSum(_.state == PendingConfirmationsReceived)
+    unconfirmed.foreach(balance =>
+      logger.trace(s"Unconfirmed balance=${balance.satoshis}"))
+    unconfirmed
+
+  }
+
+  override def getUnconfirmedBalance(
+      account: HDAccount): Future[CurrencyUnit] = {
+    val allUnspentF = spendingInfoDAO.findAllUnspent()
+    val unspentInAccountF = for {
+      allUnspent <- allUnspentF
+    } yield {
+      allUnspent.filter { utxo =>
+        HDAccount.isSameAccount(utxo.privKeyPath.path, account) &&
+        utxo.blockHash.isEmpty
+      }
+    }
+
+    unspentInAccountF.map(_.foldLeft(CurrencyUnits.zero)(_ + _.output.value))
+  }
+
+  /** Enumerates all the TX outpoints in the wallet  */
+  protected[wallet] def listOutpoints(): Future[Vector[TransactionOutPoint]] =
+    spendingInfoDAO.findAllOutpoints()
+
+  /** Gets the size of the bloom filter for this wallet  */
+  private def getBloomFilterSize(
+      pubkeys: Seq[ECPublicKey],
+      outpoints: Seq[TransactionOutPoint]): Int = {
+    // when a public key is inserted into a filter
+    // both the pubkey and the hash of the pubkey
+    // gets inserted
+    pubkeys.length * 2
+  } + outpoints.length
+
+  // todo: insert TXIDs? need to track which txids we should
+  // ask for, somehow
+  // We add all outpoints to the bloom filter as a way
+  // of working around the fact that bloom filters
+  // was never updated to incorporate SegWit changes.
+  // see this mailing list thread for context:
+  //   https://www.mail-archive.com/bitcoin-dev@lists.linuxfoundation.org/msg06950.html
+  // especially this email from Jim Posen:
+  //   https://www.mail-archive.com/bitcoin-dev@lists.linuxfoundation.org/msg06952.html
+  override def getBloomFilter(): Future[BloomFilter] = {
+    for {
+      pubkeys <- listPubkeys()
+      outpoints <- listOutpoints()
+    } yield {
+      val filterSize = getBloomFilterSize(pubkeys, outpoints)
+
+      // todo: Is this the best flag to use?
+      val bloomFlag = BloomUpdateAll
+
+      val baseBloom =
+        BloomFilter(numElements = filterSize,
+                    falsePositiveRate = walletConfig.bloomFalsePositiveRate,
+                    flags = bloomFlag)
+
+      val withPubs = pubkeys.foldLeft(baseBloom) { _.insert(_) }
+      outpoints.foldLeft(withPubs) { _.insert(_) }
+    }
+  }
+
+  override def markUTXOsAsReserved(
+      utxos: Vector[SpendingInfoDb]): Future[Vector[SpendingInfoDb]] = {
+    val updated = utxos.map(_.copyWithState(TxoState.Reserved))
+    spendingInfoDAO.updateAll(updated)
+  }
+
+  /** @inheritdoc */
+  def updateUtxoPendingStates(
+      blockHeader: BlockHeader): Future[Vector[SpendingInfoDb]] = {
+    for {
+      infos <- spendingInfoDAO.findAllPendingConfirmation
+      updatedInfos <- {
+        val updatedInfoFs =
+          infos.map(info => updateUtxoConfirmedState(info, blockHeader.hashBE))
+        Future.sequence(updatedInfoFs)
+      }
+    } yield updatedInfos
   }
 
   override def sendToAddress(
@@ -46,8 +218,7 @@ sealed abstract class Wallet extends LockedWallet with UnlockedWalletApi {
         destinations = Vector(destination),
         feeRate = feeRate,
         fromAccount = fromAccount,
-        keyManagerOpt = Some(keyManager),
-        markAsReserved = false)
+        keyManagerOpt = Some(keyManager))
       signed <- txBuilder.sign
       ourOuts <- findOurOuts(signed)
       _ <- processOurTransaction(transaction = signed,
@@ -121,7 +292,6 @@ sealed abstract class Wallet extends LockedWallet with UnlockedWalletApi {
   /** Creates a new account my reading from our account database, finding the last account,
     * and then incrementing the account index by one, and then creating that account
     *
-    * @param kmParams
     * @return
     */
   override def createNewAccount(kmParams: KeyManagerParams): Future[Wallet] = {
@@ -238,7 +408,7 @@ object Wallet extends WalletLogger {
             //probably means you haven't initialized the key manager via the
             //'CreateKeyManagerApi'
             throw new RuntimeException(
-              s"Failed to create keymanager with params=${kmParams} err=${err}")
+              s"Failed to create keymanager with params=$kmParams err=$err")
         }
 
       }
